removing typescript - just not going to work on this project
