+++ README.md
@@ -8,8 +8,5 @@ method of creating artwork using fractal lines and the `<canvas>` element.
 
 The algorithms defined here are not my own work. Full credit belongs to [RectangleWorld.com](http://rectangleworld.com)
 
-## ES6+
-> ### Please Note!
-> This package is designed for ES6+ only! It will not with versions prior to ECMA2015
 
 TODO everything
