+++ package.json
@@ -8,7 +8,8 @@
   },
   "scripts": {
     "pretest": "./node_modules/.bin/tsc",
-    "test": "./node_modules/.bin/mocha build/tests/*.spec.js",
+    "test": "./node_modules/.bin/mocha",
+    "test:watch": "./node_modules/.bin/mocha -w",
     "postinstall": "./node_modules/.bin/typings install"
   },
   "directories": {
@@ -19,7 +20,17 @@
     "generative",
     "art",
     "fractal",
-    "image"
+    "canvas",
+    "graphic",
+    "graphics",
+    "pixman",
+    "cairo",
+    "image",
+    "images",
+    "png",
+    "jpeg",
+    "jpg",
+    "gif"
   ],
   "author": {
     "name": "Matt Keeble",
@@ -29,15 +40,16 @@
     "type": "git",
     "url": "https://github.com/beldougie/generative-art.git"
   },
+  "bugs": {
+    "url": "https://github.com/beldougie/generative-art/issues"
+  },
   "license": "ISC",
   "devDependencies": {
     "chai": "^3.5.0",
-    "mocha": "^3.0.2",
-    "typescript": "^1.8.10",
-    "typings": "^1.3.3"
+    "mocha": "^3.0.2"
   },
   "dependencies": {
-    "canvas": "^1.4.0",
+    "canvas": "https://github.com/beldougie/node-canvas.git",
     "lodash": "^4.15.0"
   }
 }
