--- typings.json
@@ -1,9 +1,0 @@
-{
-  "name": "generative-art",
-  "dependencies": {},
-  "globalDependencies": {
-    "chai": "registry:dt/chai#3.4.0+20160601211834",
-    "lodash": "registry:dt/lodash#4.14.0+20160830145422",
-    "mocha": "registry:dt/mocha#2.2.5+20160720003353"
-  }
-}
